Nest test data under tests
